feat(containers): Fix parallelisation for icon
